Move keypoint detection hook to a new file.
